Created oclif subcommands and refactored repo structure
