+++ package.json
@@ -10,7 +10,7 @@
   "dependencies": {
     "@elastic/elasticsearch": "^7.5.0",
     "@oclif/command": "^1",
-    "@oclif/config": "^1",
+    "@oclif/config": "^1.14.0",
     "@oclif/plugin-help": "^2",
     "@types/fs-extra": "^8.0.1",
     "@types/js-yaml": "^3.12.1",
@@ -28,26 +28,26 @@
     "load-yaml-file": "^0.2.0",
     "lodash": "^4.17.15",
     "node-fetch": "^2.6.0",
-    "tslib": "^1",
+    "tslib": "^1.11.1",
     "uuid-by-string": "^3.0.2"
   },
   "devDependencies": {
     "@oclif/dev-cli": "^1",
     "@oclif/test": "^1",
-    "@types/chai": "^4",
-    "@types/mocha": "^5",
-    "@types/node": "^12.12.17",
-    "@typescript-eslint/eslint-plugin": "^2.11.0",
-    "@typescript-eslint/parser": "^2.11.0",
+    "@types/chai": "^4.2.11",
+    "@types/mocha": "^7.0.2",
+    "@types/node": "^13.9.2",
+    "@typescript-eslint/eslint-plugin": "^2.24.0",
+    "@typescript-eslint/parser": "^2.24.0",
     "chai": "^4",
-    "eslint": "^6.7.2",
-    "eslint-config-prettier": "^6.7.0",
-    "globby": "^10",
-    "mocha": "^5",
-    "nyc": "^13",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.10.0",
+    "globby": "^11.0.0",
+    "mocha": "^7.1.1",
+    "nyc": "^15.0.0",
     "prettier": "^1.19.1",
-    "ts-node": "^8",
-    "typescript": "^3.3"
+    "ts-node": "^8.7.0",
+    "typescript": "^3.8.3"
   },
   "engines": {
     "node": ">=8.0.0"
@@ -69,7 +69,15 @@
     "bin": "zindexer",
     "plugins": [
       "@oclif/plugin-help"
-    ]
+    ],
+    "topics": {
+      "github": {
+        "description": "Fetches data from GitHub"
+      },
+      "jira": {
+        "description": "Fetches data from Jira"
+      }
+    }
   },
   "repository": "zencrepes/zindexer",
   "scripts": {
