+++ src/commands/github/pullrequests.ts
@@ -1,22 +1,22 @@
 import { flags } from '@oclif/command';
 import cli from 'cli-ux';
 
-import Command from '../base';
-import esClient from '../utils/es/esClient';
-import esGithubLatest from '../utils/es/esGithubLatest';
-import esPushNodes from '../utils/es/esPushNodes';
-import fetchNodesUpdated from '../utils/github/fetchNodesUpdated';
-import ghClient from '../utils/github/ghClient';
+import Command from '../../base';
+import esClient from '../../utils/es/esClient';
+import esGithubLatest from '../../utils/es/esGithubLatest';
+import esPushNodes from '../../utils/es/esPushNodes';
+import fetchNodesUpdated from '../../utils/github/utils/fetchNodesUpdated';
+import ghClient from '../../utils/github/utils/ghClient';
 
-import esGetActiveSources from '../utils/es/esGetActiveSources';
-import esCheckIndex from '../utils/es/esCheckIndex';
-import ymlMappingsGPullrequests from '../schemas/gPullrequests';
+import esGetActiveSources from '../../utils/es/esGetActiveSources';
+import esCheckIndex from '../../utils/es/esCheckIndex';
 
-import { getId } from '../utils/misc/getId';
+import { getId } from '../../utils/misc/getId';
 
-import getPullrequests from '../utils/github/graphql/getPullrequests';
+import esMapping from '../../utils/github/pullrequests/esMapping';
+import fetchGql from '../../utils/github/pullrequests/fetchGql';
 
-export default class GPullrequests extends Command {
+export default class Pullrequests extends Command {
   static description =
     'Github: Fetches Pullrequests data from configured sources';
 
@@ -39,7 +39,7 @@ export default class GPullrequests extends Command {
 
     const fetchData = new fetchNodesUpdated(
       gClient,
-      getPullrequests,
+      fetchGql,
       this.log,
       userConfig.github.fetch.maxNodes,
       this.config.configDir,
@@ -69,12 +69,7 @@ export default class GPullrequests extends Command {
       cli.action.stop(' done');
 
       // Check if index exists, create it if it does not
-      await esCheckIndex(
-        eClient,
-        userConfig,
-        pullrequestsIndex,
-        ymlMappingsGPullrequests,
-      );
+      await esCheckIndex(eClient, userConfig, pullrequestsIndex, esMapping);
 
       await esPushNodes(fetchedPullrequests, pullrequestsIndex, eClient);
